move code to internal package
