+++ cmd/slice.go
@@ -1,7 +1,7 @@
 package cmd
 
 import (
-	"github.com/alexfedosov/move-tool/app"
+	"github.com/alexfedosov/move-tool/internal"
 	"github.com/spf13/cobra"
 )
 
@@ -16,7 +16,7 @@ var (
 		Short: "Slices long sample into drum rack",
 		Long:  `Slice long sample into given number of equal numberOfSlices and creates a drum rack preset`,
 		RunE: func(cmd *cobra.Command, args []string) error {
-			return app.SliceSampleIntoDrumRack(input, output, numberOfSlices, presetName)
+			return internal.SliceSampleIntoDrumRack(input, output, numberOfSlices, presetName)
 		},
 	}
 )
