+++ internal/ablmodels/drum_cell_chain.go
