+++ internal/audioutils_test.go
@@ -1,4 +1,4 @@
-package app
+package internal
 
 import (
 	"fmt"
